# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0-SNAPSHOT",
        "title": "DashApi",
        "description": "API for DashApi application",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "dashchallengesapi.com",
    "basePath": "/mobileapi",
    "tags": [
        {
            "name": "Users",
            "description": "API for users in the system"
        }
    ],
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/userapi": {
            "post": {
                "tags": [
                    "Mobile_Users"
                ],
                "description": "Create new user in system",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User that we want to create",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "New user is created",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Mobile_Users"
                ],
                "summary": "Get all users in system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Users"
                        }
                    }
                }
            },
            "patch": {
                "summary": "Update user with give ID",
                "tags": [
                    "Mobile_Users"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "update User  values of properties",
                        "schema": {
                            "$ref": "#/definitions/Update_User_Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User is updated",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/userapi/{userId}": {
            "parameters": [
                {
                    "name": "userId",
                    "in": "path",
                    "required": true,
                    "description": "ID of user that we want to find",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Mobile_Users"
                ],
                "summary": "Get user with given ID",
                "responses": {
                    "200": {
                        "description": "User is found",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete user with given ID",
                "tags": [
                    "Mobile_Users"
                ],
                "responses": {
                    "200": {
                        "description": "User is deleted",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "required": [
                "displayname",
                "username",
                "email",
                "password",
                "phoneNumber"
            ],
            "properties": {
                "displayname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                }
            }
        },
        "Users": {
            "type": "array",
            "$ref": "#/definitions/User"
        },
        "Update_User_Data": {
            "required": [
                "editUserID",
                "displayname",
                "username",
                "email",
                "password",
                "phoneNumber"
            ],
            "properties": {
                "editUserID": {
                    "type": "string"
                },
                "displayname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "friendsIds": {
                    "type": "string"
                },
                "challengesIds": {
                    "type": "string"
                },
                "AdsStatus": {
                    "type": "string"
                }
            }
        }
    }
}